修复商品查询结果未过滤淘汰、停用货品的bug；解决已结清的销售单无法查看详情的问题

+++ app/src/main/java/com/yifarj/yifadinghuobao/model/entity/CreateOrderEntity.java
@@ -39,7 +39,7 @@ public class CreateOrderEntity {
         public String DeliveryAddress;
         public String ReceiveMethod;
         public long ReceiveTime;
-        public int FinishedAmount;
+        public double FinishedAmount;
         public int CreatedUserId;
         public long CreatedTime;
         public String CreatedDevice = "";
@@ -62,7 +62,7 @@ public class CreateOrderEntity {
         public double TotalAmount;
         public float Discount;
         public double DiscountAmount;
-        public float ReBateAmount;
+        public double ReBateAmount;
         public String WarehouseName = "";
         public String LocationName = "";
         public int PriceSystemId;
